Move communication messages into a separate file
